Add vector store implementation for Azure Cognitive Search.

+++ spring-chatgpt-sample-common/src/main/java/com/microsoft/azure/spring/chatgpt/sample/common/vectorstore/DocEntry.java
@@ -16,5 +16,5 @@ public class DocEntry {
 
     private final String text;
 
-    private final List<Double> embedding;
+    private final List<Float> embedding;
 }
